Use device specific flows for editing topics
There is a known issue in Maestro and SwiftUI where an identifier can't
be added solely to the switch, rather the whole Toggle itself (label +
switch cell combinded) meaning the switch can't be activated by Maestro.

There is a workaround for our use case [2] where you seperate out the Toggle
and the toggle text label but this breaks the accessibility behaviour.

This left two options, tapOn for coordinates or indexes. Indexes should
be less brittle than coordinates but more brittle than using the
preferred identifiers.

Source [1]
> When Toggle is initialized with text, its accessibility element is a
union of a text and toggle

[1]: https://maestro.mobile.dev/platform-support/ios-swiftui#known-issues
[2]: https://github.com/alphagov/govuk-mobile-ios-app/blob/be436a44072a249c7fbcdeb69364c3b822f1d51b/GOVKit/Sources/GOVKit/SwiftUIViews/GroupedList/GroupedListRowView.swift#L110-L117